requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.83.0
h11==0.13.0
httptools==0.4.0
idna==3.3
imutils==0.5.4
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.19.1
typing-extensions==4.3.0
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.17.0
websockets==10.3
